+++ pkg/handlers/mux-handler.go
@@ -6,14 +6,14 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/cheebz/go-pub/activitypub"
-	"github.com/cheebz/go-pub/config"
-	"github.com/cheebz/go-pub/media"
-	"github.com/cheebz/go-pub/middleware"
-	"github.com/cheebz/go-pub/resources"
-	"github.com/cheebz/go-pub/responses"
-	"github.com/cheebz/go-pub/services"
-	"github.com/cheebz/go-pub/utils"
+	"github.com/cheebz/go-pub/pkg/activitypub"
+	"github.com/cheebz/go-pub/pkg/config"
+	"github.com/cheebz/go-pub/pkg/media"
+	"github.com/cheebz/go-pub/pkg/middleware"
+	"github.com/cheebz/go-pub/pkg/resources"
+	"github.com/cheebz/go-pub/pkg/responses"
+	"github.com/cheebz/go-pub/pkg/services"
+	"github.com/cheebz/go-pub/pkg/utils"
 	"github.com/cheebz/sigs"
 	"github.com/gorilla/mux"
 )
