+++ pkg/services/activitypub-service.go
@@ -7,11 +7,11 @@ import (
 	"path"
 
 	"github.com/cheebz/arb"
-	"github.com/cheebz/go-pub/activitypub"
-	"github.com/cheebz/go-pub/config"
-	"github.com/cheebz/go-pub/media"
-	"github.com/cheebz/go-pub/models"
-	"github.com/cheebz/go-pub/repositories"
+	"github.com/cheebz/go-pub/pkg/activitypub"
+	"github.com/cheebz/go-pub/pkg/config"
+	"github.com/cheebz/go-pub/pkg/media"
+	"github.com/cheebz/go-pub/pkg/models"
+	"github.com/cheebz/go-pub/pkg/repositories"
 )
 
 type ActivityPubService struct {
